Revert "feature(PAK): creating USER logic and auth implementation"
