--- src/main/java/fun/epak/pak/config/WebServiceConfig.java
@@ -1,62 +1,0 @@
-package fun.epak.pak.config;
-
-import fun.epak.pak.service.UserService;
-import org.springframework.boot.web.servlet.ServletRegistrationBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-
-import javax.servlet.annotation.WebServlet;
-
-import static org.springframework.security.config.Customizer.withDefaults;
-
-@Configuration
-@EnableWebSecurity
-public class WebServiceConfig{
-
-/*
-    @Bean
-    public UserDetailsService userDetailsService() {
-        return new UserService();
-    }
-*/
-
-    @Bean
-    protected static PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public DaoAuthenticationProvider authenticationProvider(UserService userService) {
-        DaoAuthenticationProvider authProvider = new DaoAuthenticationProvider();
-        authProvider.setUserDetailsService(userService);
-        authProvider.setPasswordEncoder(passwordEncoder());
-        return authProvider;
-    }
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http.authorizeHttpRequests((auth) -> {
-                    try {
-                        auth.antMatchers("/**").hasAuthority("USER")
-                                .and()
-                                .csrf().disable()
-                                .headers().frameOptions().disable();
-                    } catch (Exception e) {
-                        e.printStackTrace();
-                    }
-                }
-        ).httpBasic(withDefaults());
-        return http.build();
-    }
-
-
-
-
-}
