+++ src/main/java/fun/epak/pak/model/Comment.java
@@ -1,6 +1,5 @@
 package fun.epak.pak.model;
 
-import fun.epak.pak.model.user.User;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,7 +8,6 @@ import lombok.Setter;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import javax.persistence.ManyToOne;
 import java.time.LocalDate;
 
 @Entity
@@ -21,8 +19,7 @@ public class Comment {
     @Id
     @GeneratedValue
     private long id;
-    @ManyToOne
-    private User user;
+    private long userId;
     private LocalDate createDate;
     private String content;
     private long postId;
