--- src/main/java/fun/epak/pak/model/user/UserDetails.java
@@ -1,53 +1,0 @@
-package fun.epak.pak.model.user;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-
-import java.util.Collection;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-@RequiredArgsConstructor
-public class UserDetails implements org.springframework.security.core.userdetails.UserDetails {
-
-    private final User user;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        Set<UserRole> roles = user.getRoles();
-        return roles.stream()
-                .map(role -> new SimpleGrantedAuthority(role.getRole()))
-                .collect(Collectors.toList());
-    }
-
-    @Override
-    public String getPassword() {
-        return user.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return user.getEmail();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return user.getIsActive();
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return user.getIsActive();
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return user.getIsActive();
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return user.getIsActive();
-    }
-}
