--- src/main/java/fun/epak/pak/model/user/UserRole.java
@@ -1,27 +1,0 @@
-package fun.epak.pak.model.user;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.*;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserRole {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-    @ManyToOne
-    private User user;
-    private String role;
-
-    public UserRole(User user) {
-        this.user = user;
-        role = "USER";
-    }
-}
