--- src/main/java/fun/epak/pak/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package fun.epak.pak.repository;
-
-import fun.epak.pak.model.user.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-   Optional<User> findByEmail(String email);
-}
