+++ src/main/resources/application-develop.properties
@@ -1,4 +1,3 @@
-spring.jpa.defer-datasource-initialization=true
 spring.datasource.url=jdbc:h2:mem:testdb
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=pak
